database CRUD

+++ rest/handlers/product/delete_product.go
@@ -13,10 +13,14 @@ func (h *Handler) DeleteProduct(w http.ResponseWriter, r *http.Request) {
 	pId, err := strconv.Atoi(productID)
 	if err != nil {
 		fmt.Println(err)
-		util.SendError(w,http.StatusBadRequest,"Enter a valid Id")
+		util.SendError(w, http.StatusBadRequest, "Enter a valid Id")
 		return
 	}
-
+	product, _ := h.productRepo.Get(pId)
+	if product == nil {
+		util.SendError(w, 404, "Product Not Found")
+		return
+	}
 	err = h.productRepo.Delete(pId)
 	if err != nil {
 		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
